requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
cryptography==41.0.7
Pillow==10.1.0
opencv-python-headless
numpy
scipy
gunicorn
pytest
Flask-Limiter
python-dotenv
